impexpstudents/requirements.txt
Django==2.2.8
django-import-export==1.2.0
awesome-slugify==1.6.5